requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
openai>=1.40.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
